Refactor specs into standalone directory and configuration changes.
