+++ package.json
@@ -132,7 +132,7 @@
     "rollup-plugin-node-resolve": "^4.2.3",
     "staged-git-files": "^1.3.0",
     "stylus": "^0.54.8",
-    "testcafe": "^2.5.1-rc.1",
+    "testcafe": "^2.6.2",
     "testcafe-browser-provider-saucelabs": "^1.9.0"
   }
 }
