+++ test/functional/helpers/components/deserializeReactElement.js
