+++ test/functional/helpers/components/renderComponentView.jsx
