+++ test/functional/helpers/components/renderReactElement.js
