+++ test/functional/helpers/runtime/addHtmlToBody.js
