+++ test/functional/helpers/runtime/appendLaunchLibrary.js
