+++ test/functional/helpers/runtime/appendScript.js
