+++ test/functional/helpers/runtime/getReturnedEcid.js
