--- test/functional/runtime/mockVisitor/extension.json
@@ -1,26 +1,0 @@
-{
-  "displayName": "Mock Visitor",
-  "name": "adobe-mcid",
-  "platform": "web",
-  "version": "0.0.1",
-  "description": "Mock Experience Cloud ID Service for use with testing",
-  "author": {
-    "name": "adobe"
-  },
-  "viewBasePath": "src/view/",
-  "sharedModules": [
-    {
-      "name": "mcid-instance",
-      "libPath": "src/lib/sharedModules/mcidInstance.js"
-    }
-  ],
-  "configuration": {
-    "viewPath": "configuration/configuration.html",
-    "schema": {
-      "$schema": "http://json-schema.org/draft-04/schema#",
-      "type": "object",
-      "properties": {}
-    }
-  },
-  "main": "src/lib/sharedModules/mcidInstance.js"
-}
