+++ test/functional/specs/actions/applyResponse.spec.js
@@ -10,11 +10,11 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import extensionViewController from "../helpers/extensionViewController";
-import spectrum from "../helpers/spectrum";
-import testInstanceNameOptions from "../helpers/testInstanceNameOptions";
-import createExtensionViewFixture from "../helpers/createExtensionViewFixture";
-import runCommonExtensionViewTests from "../runCommonExtensionViewTests";
+import extensionViewController from "../../helpers/extensionViewController";
+import spectrum from "../../helpers/spectrum";
+import testInstanceNameOptions from "../../helpers/testInstanceNameOptions";
+import createExtensionViewFixture from "../../helpers/createExtensionViewFixture";
+import runCommonExtensionViewTests from "../../runCommonExtensionViewTests";
 
 const instanceNameField = spectrum.picker("instanceNameField");
 const responseHeadersField = spectrum.textField("responseHeadersField");
