+++ test/functional/specs/dataElements/xdmObjectFieldValidation.spec.js
