+++ test/functional/specs/runtime/redirectWithIdentity.spec.js
@@ -10,10 +10,15 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { t } from "testcafe";
-import createNetworkLogger from "./helpers/createNetworkLogger";
-import appendLaunchLibrary from "./helpers/appendLaunchLibrary";
-import { TEST_PAGE } from "./helpers/constants/url";
+import { t, ClientFunction, Selector } from "testcafe";
+import createNetworkLogger from "../../helpers/runtime/createNetworkLogger";
+import addHtmlToBody from "../../helpers/runtime/addHtmlToBody";
+import {
+  SECONDARY_TEST_PAGE,
+  TEST_PAGE
+} from "../../helpers/runtime/constants/url";
+import getReturnedEcid from "../../helpers/runtime/getReturnedEcid";
+import appendLaunchLibrary from "../../helpers/runtime/appendLaunchLibrary";
 
 const networkLogger = createNetworkLogger();
 
@@ -25,29 +30,11 @@ const container = {
         instances: [
           {
             name: "alloy",
-            edgeConfigId: "bc1a10e0-aee4-4e0e-ac5b-cdbb9abbec83:AditiTest"
+            edgeConfigId: "bc1a10e0-aee4-4e0e-ac5b-cdbb9abbec83:AditiTest",
+            thirdPartyCookiesEnabled: false
           }
         ]
       }
-    }
-  },
-  dataElements: {
-    "XDM Object 1": {
-      settings: {
-        cacheId: "47ec6bcf-a41a-4dde-8883-88c18a867d70",
-        sandbox: {
-          name: "prod"
-        },
-        schema: {
-          id:
-            "https://ns.adobe.com/unifiedjsqeonly/schemas/75bc29dc603dbb5c8ba7c9f5be97b852a48772ccc69d0921",
-          version: "1.1"
-        }
-      },
-      cleanText: false,
-      forceLowerCase: false,
-      modulePath: "adobe-alloy/dist/lib/dataElements/variable/index.js",
-      storageDuration: ""
     }
   },
   rules: [
@@ -62,21 +49,28 @@ const container = {
       ],
       actions: [
         {
-          modulePath: "adobe-alloy/dist/lib/actions/updateVariable/index.js",
-          settings: {
-            dataElementCacheId: "47ec6bcf-a41a-4dde-8883-88c18a867d70",
-            data: {
-              device: {
-                colorDepth: 42
-              }
-            }
-          }
-        },
-        {
           modulePath: "adobe-alloy/dist/lib/actions/sendEvent/index.js",
           settings: {
-            instanceName: "alloy",
-            xdm: "%XDM Object 1%"
+            instanceName: "alloy"
+          }
+        }
+      ]
+    },
+    {
+      id: "RL1653692204047",
+      name: "Append identity to urls",
+      events: [
+        {
+          modulePath: "sandbox/click.js",
+          settings: {}
+        }
+      ],
+      actions: [
+        {
+          modulePath:
+            "adobe-alloy/dist/lib/actions/redirectWithIdentity/index.js",
+          settings: {
+            instanceName: "alloy"
           }
         }
       ]
@@ -99,24 +93,39 @@ const container = {
   },
   buildInfo: {
     turbineVersion: "27.2.1",
-    turbineBuildDate: "2022-10-28T21:23:47.138Z",
-    buildDate: "2022-10-28T21:23:47.139Z",
+    turbineBuildDate: "2022-05-27T22:57:44.929Z",
+    buildDate: "2022-05-27T22:57:44.929Z",
     environment: "development"
   }
 };
 
-fixture("Update variable")
+fixture("Redirect with identity")
   .page(TEST_PAGE)
   .requestHooks([networkLogger.edgeEndpointLogs]);
 
-test("Updates a variable", async () => {
+const getLocation = ClientFunction(() => document.location.href);
+
+test("Redirects with an identity", async () => {
   await appendLaunchLibrary(container);
+
+  await addHtmlToBody(
+    `<a href="${SECONDARY_TEST_PAGE}"><div id="mylink">My link</div></a>`
+  );
   // The requestLogger.count method uses TestCafe's smart query
   // assertion mechanism, so it will wait for the request to be
   // made or a timeout is reached.
   await t.expect(networkLogger.edgeEndpointLogs.count(() => true)).eql(1);
-  const requestBody = JSON.parse(
-    networkLogger.edgeEndpointLogs.requests[0].request.body
+  await t.click(Selector("#mylink"));
+  await t.expect(getLocation()).contains(SECONDARY_TEST_PAGE);
+  await appendLaunchLibrary(container);
+
+  // Events are: page load, link click, page load.
+  await t.expect(networkLogger.edgeEndpointLogs.count(() => true)).eql(3);
+  const pageLoad1Ecid = getReturnedEcid(
+    networkLogger.edgeEndpointLogs.requests[0]
   );
-  await t.expect(requestBody.events[0].xdm.device.colorDepth).eql(42);
+  const pageLoad2Ecid = getReturnedEcid(
+    networkLogger.edgeEndpointLogs.requests[2]
+  );
+  await t.expect(pageLoad1Ecid).eql(pageLoad2Ecid);
 });
