+++ test/functional/specs/runtime/visitorMigration.spec.js
@@ -10,12 +10,11 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { t, ClientFunction, Selector } from "testcafe";
-import createNetworkLogger from "./helpers/createNetworkLogger";
-import addHtmlToBody from "./helpers/addHtmlToBody";
-import { SECONDARY_TEST_PAGE, TEST_PAGE } from "./helpers/constants/url";
-import getReturnedEcid from "./helpers/getReturnedEcid";
-import appendLaunchLibrary from "./helpers/appendLaunchLibrary";
+import { t } from "testcafe";
+import createNetworkLogger from "../../helpers/runtime/createNetworkLogger";
+import appendLaunchLibrary from "../../helpers/runtime/appendLaunchLibrary";
+import getReturnedEcid from "../../helpers/runtime/getReturnedEcid";
+import { TEST_PAGE } from "../../helpers/runtime/constants/url";
 
 const networkLogger = createNetworkLogger();
 
@@ -27,11 +26,14 @@ const container = {
         instances: [
           {
             name: "alloy",
-            edgeConfigId: "bc1a10e0-aee4-4e0e-ac5b-cdbb9abbec83:AditiTest",
-            thirdPartyCookiesEnabled: false
+            edgeConfigId: "bc1a10e0-aee4-4e0e-ac5b-cdbb9abbec83:AditiTest"
           }
         ]
       }
+    },
+    "adobe-mcid": {
+      displayName: "Mock Visitor",
+      settings: {}
     }
   },
   rules: [
@@ -47,25 +49,6 @@ const container = {
       actions: [
         {
           modulePath: "adobe-alloy/dist/lib/actions/sendEvent/index.js",
-          settings: {
-            instanceName: "alloy"
-          }
-        }
-      ]
-    },
-    {
-      id: "RL1653692204047",
-      name: "Append identity to urls",
-      events: [
-        {
-          modulePath: "sandbox/click.js",
-          settings: {}
-        }
-      ],
-      actions: [
-        {
-          modulePath:
-            "adobe-alloy/dist/lib/actions/redirectWithIdentity/index.js",
           settings: {
             instanceName: "alloy"
           }
@@ -90,39 +73,23 @@ const container = {
   },
   buildInfo: {
     turbineVersion: "27.2.1",
-    turbineBuildDate: "2022-05-27T22:57:44.929Z",
-    buildDate: "2022-05-27T22:57:44.929Z",
+    turbineBuildDate: "2022-04-29T16:01:37.616Z",
+    buildDate: "2022-04-29T16:01:37.616Z",
     environment: "development"
   }
 };
 
-fixture("Redirect with identity")
+fixture("Visitor migration")
   .page(TEST_PAGE)
   .requestHooks([networkLogger.edgeEndpointLogs]);
 
-const getLocation = ClientFunction(() => document.location.href);
-
-test("Redirects with an identity", async () => {
+test("waits for Visitor to be initialized before running", async () => {
   await appendLaunchLibrary(container);
-
-  await addHtmlToBody(
-    `<a href="${SECONDARY_TEST_PAGE}"><div id="mylink">My link</div></a>`
-  );
   // The requestLogger.count method uses TestCafe's smart query
   // assertion mechanism, so it will wait for the request to be
   // made or a timeout is reached.
   await t.expect(networkLogger.edgeEndpointLogs.count(() => true)).eql(1);
-  await t.click(Selector("#mylink"));
-  await t.expect(getLocation()).contains(SECONDARY_TEST_PAGE);
-  await appendLaunchLibrary(container);
-
-  // Events are: page load, link click, page load.
-  await t.expect(networkLogger.edgeEndpointLogs.count(() => true)).eql(3);
-  const pageLoad1Ecid = getReturnedEcid(
-    networkLogger.edgeEndpointLogs.requests[0]
-  );
-  const pageLoad2Ecid = getReturnedEcid(
-    networkLogger.edgeEndpointLogs.requests[2]
-  );
-  await t.expect(pageLoad1Ecid).eql(pageLoad2Ecid);
+  const ecid = getReturnedEcid(networkLogger.edgeEndpointLogs.requests[0]);
+  // This is the ID returned from the mock visitor extension.
+  await t.expect(ecid).eql("00781847927133700121980094732316198575");
 });
